ref: Rename components of filters more accurately
