+++ src/modules/catalogue/components/CatalogueFilter/CatalogueFilterSection.styled.ts
