+++ src/modules/catalogue/components/CatalogueFilter/CatalogueFilterSectionOption.styled.ts
